requirements.txt
TableauScraper
tweepy
pillow